Make cl a feature
