+++ libindy-crypto/src/cl/issuer.rs
@@ -4,8 +4,8 @@ use errors::IndyCryptoError;
 use pair::*;
 use cl::constants::*;
 use cl::helpers::*;
-use utils::commitment::*;
-use utils::get_hash_as_int;
+use cl::commitment::get_pedersen_commitment;
+use cl::hash::get_hash_as_int;
 
 use std::collections::{HashMap, HashSet};
