+++ libindy-crypto/src/cl/logger.rs
